Bug fix and update to Kayenta.
BUG FIX:
In Source/Payette_iterative_solvers.py, the new guess for depsdt was
never saved in the Newton routine so that update state never got it
and was always using the old value of depsdt.  For small stresses, Newton
could still converge, but later always jumped in to the simplex routine.

KAYENTA UPDATE:
In preparation for Kayenta to be put in to Lambda, there needs to be only
one version distributed to CTH, ALEGRA, and Sierra SM.  Since CTH and 
Sierra SM both use the orthotropic joint option, I made the orthotropic
joint option the default in Kayenta.  The kayenta interface and signature 
files reflect that change.  ***BUT*** the version of Kayenta in the trunk
does not since I would like to speak to Rebecca Brannon about this change
before I commit to the trunk.  So, until then, use branches/tjfulle/ortho
in place of trunk/src for the Kayenta material model.