Big change to how input files are parsed.
DETAILS
Previously at startup, only the simulation blocks were parsed from
the input file.  The rest of the blocks were not parsed, but just read
in to list and stored with the appropriate simulation block.  Now,
at startup, all blocks are read and put in to dictionaries of their own.
This allows for better checking of input file formatting and is more
easily adapted to new features (like the optimization).

The change required that all files that used the input file list be
modified to use the new dictionary format.