Hopefully the Lambda Mie Gruneisen Stuff works now
The Lambda MieGruneisen stuff should work now. It accepts similar
syntax to that of the Lambda Driver, but it is all in the material
block and not the legs block.

Getting Payette to drive EOS models will be quite a pain because
Payette has been built up thinking that it would always drive through
time. Some serious refactoring will have to take place.

Also adding more diagnostic output on building Payette