New flag to testPayette "-u".
DETAILS:
What the -u flag does is use the Toolset/__found_tests__.py
file to find the tests.  __found_tests__.py is a file that is generated
automagically by testPayette and is just a dump of the "conforming" list
of tests found by find_tests.  Using the -u flag, instead of using
find_tests to search through the different test directories to find
conforming tests, the conforming tests from the last time find_tests
was called are loaded from the file.  __found_tests__.py is a pickle dump
and is ignored by the repository.