Added Payette_input_parser.py with associated InputParser class.
Previously, the user input was read and put in a pretty cryptic dictionary.
To get any information, one had to know the structure of the dictionary -
not very good.  Now, the user input is read in to a InputParser class
which has methods to access the data -> much better.

All files that somehow used the raw user input were modified, except Payette_barf.py.
All built in tests pass.

barf processing does not yet work.