Enhanced PayetteError exception raising and handling.
Moved PayetteError to the Payette_container.py class file and
added try: ... except PayetteError: ... blocks around calls to
run_job in Payette_run so that if multiple simulations are being run
a crashed job will not stop all.

If the --debug flag is sent to runPayette the simulations will stop if
any crash.