Fixed the Hugoniot discrepency
Found that assigning a variable from a numpy
array gives you a reference to that array, so if the
value in the array changes, the value in the variable
also changes (in my case it happened to be the variable
holding the initial energy).

The hugoniot is now completely in agreement with the
one produced by Bill Scherzinger's EOS driver.