Cleaned up the eos_driver() routine.
Made the eos_driver() routine so that it uses the
store_data() and get_data() functions for everything needed.

Also ensured that when store_data is called for a scalar, it will
be converted to a float so as to catch instances when the scalar
being stored happens to be a numpy array with one value.

All fast benchmarks pass (except kayenta-multileg-biaxial-plane-strain).