Extensive unit implementation.
Units MUST be defined when registering ALL parameters.
When registering data, they may be given, but are not yet
required.

ALL EOS models are required to have ALL parameters and registered
data with valid units (as well as input units and output units
defined in the boundary block of the input file).

40 of 41 benchmarks pass (some Kayenta benchmark fails) and the
Mie Gruneisen outputs for copper match exactly the outputs from
Bill Scherzinger's eos driver.