Adjust fail tolerance on tests if running on architecture other than what was used to create the gold file.
With this change, all Kayenta tests pass on darwin, and all but 5 fast tests
psss on RHEL5.  The 5 that don't pass diff.

This change will require adding a self.arch = "architecture" to each test when
created.  The default is "darwin" since all of the tests were baselined on darwin.
I'll go through tests and add the architecture explicitly to each test, remove the
default, and put in a check that the architecture is defined for all new tests later.