Added two new features
1) Preprocessing

add a

begin preprocessing
  some_value = x
end preprocessing

block in your input file and anywhere you can have

param = {some_value}

and some_value will be put there.

2) Pressure leg

Defining a single stress value in a deformation leg will be interpreted
as a pressure leg (positive in compression).  i.e., the stress
components will be set to

-p -p -p 0 0 0

if p is the value in the input file.