Added new interface to optimization.
OLD INTERFACE
With the old interface to the optimization, the user gives a list
of variables to optimize, a gold file and a list of variables to minimize.
After a simulation, the maximum normalized RMS error between the gold and output
files for the minimization variables is computed and returned to the optimizer.

NEW INTERFACE
Instead of specifying variables to minimize and a gold file, the user specifies
the path to a python file that contains the objective function and returns the
the error:

def obj_fn(*args)
    sim_output_file_path = args[0]
    [code]
    return error

It is up to the user to determine the error.  This method is infinitely more
flexible and allows the user to optimize in ways that cannot be prepackaged
as before.

In the material block of the input file, the parameters to be optimized must be
specified like:

K = {some_name_for_k}
G = {some_name_for_g}

Then, in the optimization block, we optimize as follows

optimize some_name_for_k ...
optimize some_name_for_g ...

and the optimizer substitutes the name in the optimization block
with those in the {} in the material block.  Clear?

RATIONAL FOR NEW INTERFACE
The old interface carried a lot of assumption, the biggest being that the data in the
gold file and the output file have the same length.  It works great for toy problems,
but when doing a real optimization it is not flexible enough.