New configuration process.
These changes should be transparent to users, but the configure process
was cleaned up and changed a bit.

The big change is the use of a new "DOTPAYETTE" directory.

If the user has DOTPAYETTE defined in their environment, then that
directory is used for putting configuration info.  If it is not
defined, a directory is created (~/.payette) and then used.

Some paths were changed around, that is why so many files were
touched.

All tests pass.