Added overlay plots.
In the lower left corner of the visualization window there are now
two buttons.  One button to open an overlay file and one to close an open
overlay file.  The list below the two buttons shows all currently
open overlay files.  Any data in the overlay files that matches in column
name the currently displayed Payette data will be plotted automatically
in the plot window.  Any column names in the overlay file that do not
match the currently displayed Payette data will be ignored.  Also, time
data labels are not drawn on the overlay plots.