preprocessor is MUCH more powerful now
The preprocessor is essentially a big block of python
that can be executed to create new variables (except
you can't do list comprehensions, logic, or other super
fun things). But you can use things like abs, sin, cos, pi
and variables defined in previous lines.

It works flawlessly but the implementation leaves something
to be desired. It's clean enough, but nothing to be proud of.