Preprocessor name checking and print out values
The preprocessor now looks for variable names that may be
ambiguous (like nested names 'bar' and 'foo_bar_baz') and
errors out if they exist.

Also, the preprocessor prints out all the values to the terminal
when it finishes its work (to aid in debugging). I figure that
most of the time there will be so few values that it is
inconsequential and that if you have a lot of values then you
will want to check that they make sense.

The runExamples script referenced runPayette but has been changed
to use just 'payette'.