'expect_constant' entry in benchmark py files
This change adds a 'self.expect_constant' variable for use in the
benchmark .py files. An example of its use is

self.expect_constant = ['I1', 'PRESSURE', 'EVOL']

This uses a new method of comparing the .out to the .gold for the
variables listed. It looks at averages and standard deviations instead
of a straight RMS. This makes it so that random noise in problems (like
in I1, PRESSURE, and EVOL in pure shear problems) does not cause the
test to fail just because you have a slightly different compiler,
version of python, or a different platform. That being said - all
expected tests now pass on Linux (in fact, all ortho tests pass even)!
I am using EPD_free 7.3-2 64-bit with gfortran version 4.6.3.